refactor(agor-ui): move state ownership into zustand store (Phase 4 PR2)
Move useAgorData's state ownership from its in-hook `useState<DataMaps>` into
the PR1 zustand store, keeping the hook's return signature byte-for-byte
identical so App.tsx, the prop-drilling chain, and the context providers are
untouched. The hook becomes the sole DRIVER of the store: the fetch effect and
socket subscriptions dispatch store actions; the hook reads full state back via
`useStore` and returns the same `UseAgorDataResult`.

Decomposed (not monolithic):
- store/agorMaps.ts        — DataMaps shape + pure index/merge reducers
                             (incl. Phase-1.5 buildSessionMaps), MAP_KEYS/pickMaps
- store/agorHydration.ts   — Phase-1.5 hydration ported VERBATIM: per-collection
                             liveRevisions counters, monotonic hydrationGeneration
                             tokens, bumpRevision, runHydration (snapshot-before /
                             recheck-after, apply-on-quiet else discard+retry —
                             immediate x4 then capped-exponential backoff, never
                             give up until quiet/cancelled). Non-reactive module
                             state (hot-path bumps must not notify subscribers).
- store/agorStore.ts       — vanilla store + primitives (setMap/applyMaps/
                             replaceMaps/evictBranchAndSessions immer cascade/
                             reset/resetMaps/meta setters)
- store/agorRealtimeActions.ts — socket handlers writing through store primitives,
                             each bumping its per-collection revision (cascade
                             branch-eviction bumps sessions too)

immer breadth/depth rule: cascade/multi-map via immer (evictBranchAndSessions)
and applyMaps; hot single-entity *:patched via raw setMap reducers.

Gate: the new Phase-1.5 useAgorData.test.tsx (incl. skip-apply-on-race +
per-collection bump + bulk-write revision-bump cases) passes UNEDITED.
agor-ui typecheck + lint + test (525) + build all green.
