+++ apps/agor-ui/src/hooks/useAgorData.ts
@@ -2,19 +2,28 @@
 /**
  * React hook for fetching and subscribing to Agor data
  *
- * Manages sessions, tasks, boards with real-time WebSocket updates
+ * Manages sessions, tasks, boards with real-time WebSocket updates.
+ *
+ * Phase 4, PR2: state OWNERSHIP now lives in the zustand store (`agorStore`).
+ * This hook is the single DRIVER of that store — the fetch effect and socket
+ * subscriptions dispatch store actions, and the hook reads full state back via
+ * `useStore` and returns the SAME `UseAgorDataResult`. `App.tsx`, the
+ * prop-drilling chain, and the context providers are byte-for-byte unaffected;
+ * later PRs peel consumers onto narrow selector subscriptions. The realtime
+ * entity reducers + index/merge helpers moved to `../store/agorRealtimeActions`
+ * and `../store/agorMaps`, and the Phase-1.5 hydration bookkeeping
+ * (per-collection revision counters, generation tokens, `runHydration`) to
+ * `../store/agorHydration` — all ported VERBATIM. See
+ * `~/.claude/plans/zustand-migration.md`.
  */
 
 import type {
   AgorClient,
-  Artifact,
   Board,
   BoardComment,
-  BoardEntityObject,
   Branch,
   CardType,
   CardWithType,
-  GatewayChannel,
   MCPServer,
   Repo,
   Session,
@@ -22,14 +31,39 @@ import type {
 } from '@agor-live/client';
 import { ENTITY_PATH_SEGMENTS, findByShortIdPrefix, PAGINATION } from '@agor-live/client';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useStore } from 'zustand';
+import {
+  bumpFirstPaintMergeRevisions,
+  bumpRevision,
+  cancelAllHydrations,
+  cancelAndFailAllHydrations,
+  resetHydrationRevisions,
+  runHydration,
+} from '../store/agorHydration';
+import {
+  buildBoardObjectMaps,
+  buildById,
+  buildSessionMaps,
+  buildSessionMcpMap,
+  type DataMaps,
+  pickMaps,
+  replaceIfChanged,
+} from '../store/agorMaps';
+import * as realtime from '../store/agorRealtimeActions';
+import { agorStore } from '../store/agorStore';
 import { createInitialLoadDebugTimer, isInitialLoadDebugEnabled } from '../utils/initialLoadDebug';
-import { shallowEqualEntity } from '../utils/shallowEqual';
 import { TOKENS_REFRESHED_EVENT } from '../utils/singleFlightRefresh';
 import {
   resolveBoardFromUrlPure,
   resolveBranchFromShortIdPure,
   resolveSessionFromShortIdPure,
 } from '../utils/urlResolution';
+
+export type { DataMaps } from '../store/agorMaps';
+// Re-exported for backward compatibility: the canonical data-map shape +
+// empties now live in `../store/agorMaps`, but existing importers (and tests)
+// reference them via this module.
+export { EMPTY_MAPS } from '../store/agorMaps';
 
 // Canonical list of initial-load items tracked by the loading checklist —
 // the ESSENTIAL set the first-paint gate blocks on. Internal only; consumers
@@ -57,47 +91,6 @@ const INITIAL_LOAD_ITEMS = [
 
 export type InitialLoadItemKey = (typeof INITIAL_LOAD_ITEMS)[number]['key'];
 
-// Skip-apply-on-race background hydration retry schedule. A hydration applies
-// its full-set snapshot ONLY if no live write to the target collection(s)
-// raced the fetch (proven via the per-collection `liveRevisionsRef` counters);
-// if one did, the snapshot is DISCARDED and refetched from a fresh revision
-// baseline — never overlaid/reconciled.
-//
-// It retries UNTIL it lands a quiet window, and NEVER gives up: skipping the
-// apply forever would leave Home empty/incomplete indefinitely on a busy
-// workspace, because live subscriptions deliver only CHANGES, not a backfill of
-// existing rows (board switching doesn't refetch, and a reconnect may never
-// fire). The first few retries are immediate (the race window is ~one fetch RTT,
-// so a single transient race converges instantly), then capped exponential
-// backoff lets a sustained live-write burst settle without busy-looping. Each
-// retry RE-snapshots the revision and RE-fetches; a racy snapshot is never
-// force-applied. Per-collection quiet windows are short, so this converges fast
-// (branches almost immediately; sessions once their write churn quiets).
-//
-// Delays PRECEDE the attempt they guard (the delay for attempt N runs before
-// fetch N, not after it). Loops are cancelled — not abandoned mid-flight — via
-// the per-collection generation tokens (`hydrationGenerationRef`): a newer
-// hydration (reconnect) or an unmount/reset supersedes older loops so they stop
-// retrying and never apply a stale snapshot or leak a timer.
-const HYDRATION_IMMEDIATE_RETRIES = 4;
-const HYDRATION_BACKOFF_BASE_MS = 200;
-const HYDRATION_BACKOFF_CAP_MS = 5000;
-
-// Hydrated collections that the background hydration replaces wholesale. Each
-// has its own live-write revision counter (`liveRevisionsRef`) so a write to
-// one collection never blocks another's hydration from applying.
-type HydratedCollection =
-  | 'sessions'
-  | 'branches'
-  | 'boardObjects'
-  | 'cards'
-  | 'comments'
-  | 'mcpServers'
-  | 'sessionMcp'
-  | 'gatewayChannels'
-  | 'artifacts'
-  | 'oauth';
-
 // First-paint bound for the global (non-board-scoped) sessions slice. Covers
 // Home's "My Sessions" + "Team activity" feeds (both show only recent items)
 // and seeds enough of `sessionById` to resolve `/s/<id>` deep links. The FULL
@@ -119,59 +112,6 @@ export interface InitialLoadItem {
 
 export type InitialLoadingStage = 'idle' | 'fetching' | 'indexing';
 
-/**
- * All server-backed data maps held in a single state object.
- *
- * Adding a new map here + to EMPTY_MAPS is all that's required —
- * `setMaps(EMPTY_MAPS)` in the reset effect covers every field automatically.
- */
-export type DataMaps = {
-  sessionById: Map<string, Session>;
-  sessionsByBranch: Map<string, Session[]>;
-  boardById: Map<string, Board>;
-  boardObjectById: Map<string, BoardEntityObject>;
-  boardObjectsByBoardId: Map<string, BoardEntityObject[]>;
-  // Global placement lookup. Branch placements are unique because a branch can
-  // only have one board-object row at a time.
-  boardObjectByBranchId: Map<string, BoardEntityObject>;
-  // Global placement lookup. Cards follow the same one-row-per-card service
-  // contract as branches; callers needing board-scoped iteration should use
-  // boardObjectsByBoardId instead.
-  boardObjectByCardId: Map<string, BoardEntityObject>;
-  commentById: Map<string, BoardComment>;
-  cardById: Map<string, CardWithType>;
-  cardTypeById: Map<string, CardType>;
-  repoById: Map<string, Repo>;
-  branchById: Map<string, Branch>;
-  userById: Map<string, User>;
-  mcpServerById: Map<string, MCPServer>;
-  gatewayChannelById: Map<string, GatewayChannel>;
-  artifactById: Map<string, Artifact>;
-  sessionMcpServerIds: Map<string, string[]>;
-  userAuthenticatedMcpServerIds: Set<string>;
-};
-
-export const EMPTY_MAPS: DataMaps = {
-  sessionById: new Map(),
-  sessionsByBranch: new Map(),
-  boardById: new Map(),
-  boardObjectById: new Map(),
-  boardObjectsByBoardId: new Map(),
-  boardObjectByBranchId: new Map(),
-  boardObjectByCardId: new Map(),
-  commentById: new Map(),
-  cardById: new Map(),
-  cardTypeById: new Map(),
-  repoById: new Map(),
-  branchById: new Map(),
-  userById: new Map(),
-  mcpServerById: new Map(),
-  gatewayChannelById: new Map(),
-  artifactById: new Map(),
-  sessionMcpServerIds: new Map(),
-  userAuthenticatedMcpServerIds: new Set(),
-};
-
 interface UseAgorDataResult extends DataMaps {
   initialLoadItems: InitialLoadItem[];
   initialLoadComplete: boolean;
@@ -179,128 +119,6 @@ interface UseAgorDataResult extends DataMaps {
   loading: boolean;
   error: string | null;
   refetch: () => Promise<void>;
-}
-
-// Generic byId-map replacer used by the per-entity `*Patched` handlers below.
-// Returns `prev` unchanged when the incoming entity is shallow-equal to what
-// we already hold — combined with the wrapper-level no-op short-circuit in
-// `setMapSlice`, idempotent server-side patches become true no-ops. The
-// per-entity handlers stay responsible for archive / branch-migration /
-// cross-map cleanup; this helper only covers the plain "replace one entry"
-// case.
-function replaceIfChanged<T extends object>(
-  prev: Map<string, T>,
-  id: string,
-  entity: T
-): Map<string, T> {
-  const existing = prev.get(id);
-  if (existing && shallowEqualEntity(existing, entity)) return prev;
-  const next = new Map(prev);
-  next.set(id, entity);
-  return next;
-}
-
-// Build a plain `byId` Map from a fetched list. Used by the background
-// (non-gated) fetches whose results land via their own setter rather than the
-// single atomic setMaps the essential gate performs.
-function buildById<T>(list: readonly T[], key: keyof T): Map<string, T> {
-  const map = new Map<string, T>();
-  for (const item of list) {
-    map.set(item[key] as unknown as string, item);
-  }
-  return map;
-}
-
-// Group session-MCP relationship rows by session_id.
-function buildSessionMcpMap(
-  list: readonly { session_id: string; mcp_server_id: string }[]
-): Map<string, string[]> {
-  const map = new Map<string, string[]>();
-  for (const relationship of list) {
-    const ids = map.get(relationship.session_id);
-    if (ids) ids.push(relationship.mcp_server_id);
-    else map.set(relationship.session_id, [relationship.mcp_server_id]);
-  }
-  return map;
-}
-
-// Derived board-object index set, built once from a fetched list. Shared by
-// the essential (board-scoped, first-paint) index build and the background
-// full-hydration pass — single source of truth so the two can't diverge.
-function buildBoardObjectMaps(list: readonly BoardEntityObject[]): {
-  boardObjectById: Map<string, BoardEntityObject>;
-  boardObjectsByBoardId: Map<string, BoardEntityObject[]>;
-  boardObjectByBranchId: Map<string, BoardEntityObject>;
-  boardObjectByCardId: Map<string, BoardEntityObject>;
-} {
-  const boardObjectById = new Map<string, BoardEntityObject>();
-  const boardObjectsByBoardId = new Map<string, BoardEntityObject[]>();
-  const boardObjectByBranchId = new Map<string, BoardEntityObject>();
-  const boardObjectByCardId = new Map<string, BoardEntityObject>();
-  for (const boardObject of list) {
-    boardObjectById.set(boardObject.object_id, boardObject);
-
-    const bucket = boardObjectsByBoardId.get(boardObject.board_id);
-    if (bucket) bucket.push(boardObject);
-    else boardObjectsByBoardId.set(boardObject.board_id, [boardObject]);
-
-    if (boardObject.branch_id) {
-      boardObjectByBranchId.set(boardObject.branch_id, boardObject);
-    }
-    if (boardObject.card_id) {
-      boardObjectByCardId.set(boardObject.card_id, boardObject);
-    }
-  }
-  return { boardObjectById, boardObjectsByBoardId, boardObjectByBranchId, boardObjectByCardId };
-}
-
-// Build the session lookups (`sessionById` + branch-bucketed `sessionsByBranch`)
-// from a flat session list. Shared by the bounded first-paint build and the
-// background full-hydration pass so the two can't diverge. Mirrors the realtime
-// handlers: archived sessions stay in `sessionById` (so a direct archived-link
-// can open the drawer) but are kept OUT of the branch buckets (so they never
-// reappear as branch/board cards). Cross-branch remote-created sessions are
-// projected as muted surrogate children under the creating session's branch.
-function buildSessionMaps(sessionsList: readonly Session[]): {
-  sessionById: Map<string, Session>;
-  sessionsByBranch: Map<string, Session[]>;
-} {
-  const sessionsById = new Map<string, Session>();
-  const sessionsByBranchId = new Map<string, Session[]>();
-
-  for (const session of sessionsList) {
-    sessionsById.set(session.session_id, session);
-    if (session.archived) continue;
-    const branchId = session.branch_id;
-    if (!sessionsByBranchId.has(branchId)) sessionsByBranchId.set(branchId, []);
-    sessionsByBranchId.get(branchId)!.push(session);
-  }
-
-  for (const sourceSession of sessionsList) {
-    if (sourceSession.archived) continue;
-    for (const relationship of sourceSession.remote_relationships?.as_source ?? []) {
-      if (relationship.relationship_type !== 'remote_create') continue;
-
-      const targetSession = sessionsById.get(relationship.target_session_id);
-      if (!targetSession) continue;
-
-      const sourceBranchSessions = sessionsByBranchId.get(sourceSession.branch_id) ?? [];
-      if (sourceBranchSessions.some((session) => session.session_id === targetSession.session_id)) {
-        continue;
-      }
-
-      const remoteSurrogate = createRemoteSurrogateSession(
-        sourceSession,
-        targetSession,
-        relationship
-      );
-      if (!remoteSurrogate) continue;
-
-      sessionsByBranchId.set(sourceSession.branch_id, [...sourceBranchSessions, remoteSurrogate]);
-    }
-  }
-
-  return { sessionById: sessionsById, sessionsByBranch: sessionsByBranchId };
 }
 
 // Parse the leading entity segment out of the current pathname, e.g.
@@ -374,130 +192,6 @@ function resolveDisplayedBoardId(
   }
 }
 
-function removeBoardObjectFromBoardBucket(
-  buckets: Map<string, BoardEntityObject[]>,
-  boardObject: BoardEntityObject
-): Map<string, BoardEntityObject[]> {
-  const bucket = buckets.get(boardObject.board_id);
-  if (!bucket?.some((item) => item.object_id === boardObject.object_id)) return buckets;
-
-  const next = new Map(buckets);
-  const filtered = bucket.filter((item) => item.object_id !== boardObject.object_id);
-  if (filtered.length > 0) next.set(boardObject.board_id, filtered);
-  else next.delete(boardObject.board_id);
-  return next;
-}
-
-function upsertBoardObjectInMaps(
-  prev: DataMaps,
-  boardObject: BoardEntityObject,
-  mode: 'create' | 'patch'
-): DataMaps {
-  const existing = prev.boardObjectById.get(boardObject.object_id);
-  if (mode === 'create' && existing) return prev;
-  if (mode === 'patch' && existing && shallowEqualEntity(existing, boardObject)) return prev;
-
-  const boardObjectById = new Map(prev.boardObjectById);
-  boardObjectById.set(boardObject.object_id, boardObject);
-
-  let boardObjectsByBoardId = prev.boardObjectsByBoardId;
-  if (existing && existing.board_id !== boardObject.board_id) {
-    boardObjectsByBoardId = removeBoardObjectFromBoardBucket(boardObjectsByBoardId, existing);
-  }
-
-  const bucket = boardObjectsByBoardId.get(boardObject.board_id) ?? [];
-  const bucketIndex = bucket.findIndex((item) => item.object_id === boardObject.object_id);
-  if (
-    bucketIndex === -1 ||
-    bucket[bucketIndex] !== boardObject ||
-    !shallowEqualEntity(bucket[bucketIndex], boardObject)
-  ) {
-    const nextBuckets = new Map(boardObjectsByBoardId);
-    if (bucketIndex === -1) {
-      nextBuckets.set(boardObject.board_id, [...bucket, boardObject]);
-    } else {
-      const updatedBucket = [...bucket];
-      updatedBucket[bucketIndex] = boardObject;
-      nextBuckets.set(boardObject.board_id, updatedBucket);
-    }
-    boardObjectsByBoardId = nextBuckets;
-  }
-
-  let boardObjectByBranchId = prev.boardObjectByBranchId;
-  if (existing?.branch_id && existing.branch_id !== boardObject.branch_id) {
-    boardObjectByBranchId = new Map(boardObjectByBranchId);
-    boardObjectByBranchId.delete(existing.branch_id);
-  }
-  if (boardObject.branch_id) {
-    const existingByBranch = boardObjectByBranchId.get(boardObject.branch_id);
-    if (!existingByBranch || !shallowEqualEntity(existingByBranch, boardObject)) {
-      boardObjectByBranchId =
-        boardObjectByBranchId === prev.boardObjectByBranchId
-          ? new Map(boardObjectByBranchId)
-          : boardObjectByBranchId;
-      boardObjectByBranchId.set(boardObject.branch_id, boardObject);
-    }
-  }
-
-  let boardObjectByCardId = prev.boardObjectByCardId;
-  if (existing?.card_id && existing.card_id !== boardObject.card_id) {
-    boardObjectByCardId = new Map(boardObjectByCardId);
-    boardObjectByCardId.delete(existing.card_id);
-  }
-  if (boardObject.card_id) {
-    const existingByCard = boardObjectByCardId.get(boardObject.card_id);
-    if (!existingByCard || !shallowEqualEntity(existingByCard, boardObject)) {
-      boardObjectByCardId =
-        boardObjectByCardId === prev.boardObjectByCardId
-          ? new Map(boardObjectByCardId)
-          : boardObjectByCardId;
-      boardObjectByCardId.set(boardObject.card_id, boardObject);
-    }
-  }
-
-  return {
-    ...prev,
-    boardObjectById,
-    boardObjectsByBoardId,
-    boardObjectByBranchId,
-    boardObjectByCardId,
-  };
-}
-
-function removeBoardObjectFromMaps(prev: DataMaps, boardObject: BoardEntityObject): DataMaps {
-  const existing = prev.boardObjectById.get(boardObject.object_id);
-  if (!existing) return prev;
-
-  const boardObjectById = new Map(prev.boardObjectById);
-  boardObjectById.delete(existing.object_id);
-
-  let boardObjectByBranchId = prev.boardObjectByBranchId;
-  if (
-    existing.branch_id &&
-    boardObjectByBranchId.get(existing.branch_id)?.object_id === existing.object_id
-  ) {
-    boardObjectByBranchId = new Map(boardObjectByBranchId);
-    boardObjectByBranchId.delete(existing.branch_id);
-  }
-
-  let boardObjectByCardId = prev.boardObjectByCardId;
-  if (
-    existing.card_id &&
-    boardObjectByCardId.get(existing.card_id)?.object_id === existing.object_id
-  ) {
-    boardObjectByCardId = new Map(boardObjectByCardId);
-    boardObjectByCardId.delete(existing.card_id);
-  }
-
-  return {
-    ...prev,
-    boardObjectById,
-    boardObjectsByBoardId: removeBoardObjectFromBoardBucket(prev.boardObjectsByBoardId, existing),
-    boardObjectByBranchId,
-    boardObjectByCardId,
-  };
-}
-
 function hasIdMatchingPrefix<T>(
   prefix: string,
   entries: Iterable<T>,
@@ -522,111 +216,34 @@ function hasIdMatchingPrefix<T>(
  *                                  active-list query omits it because it is archived, fetch it by ID.
  * @returns Sessions, boards, loading state, and refetch function
  */
-
-function preserveSessionRelationshipFields(session: Session, existing?: Session): Session {
-  if (!existing) return session;
-
-  const remoteRelationships = session.remote_relationships ?? existing.remote_relationships;
-  const remoteSurrogate = session.remote_surrogate ?? existing.remote_surrogate;
-
-  if (
-    remoteRelationships === session.remote_relationships &&
-    remoteSurrogate === session.remote_surrogate
-  ) {
-    return session;
-  }
-
-  return {
-    ...session,
-    ...(remoteRelationships !== undefined && { remote_relationships: remoteRelationships }),
-    ...(remoteSurrogate !== undefined && { remote_surrogate: remoteSurrogate }),
-  };
-}
-
-function createRemoteSurrogateSession(
-  sourceSession: Session,
-  targetSession: Session,
-  relationship: NonNullable<NonNullable<Session['remote_relationships']>['as_source']>[number]
-): Session | null {
-  if (relationship.relationship_type !== 'remote_create') return null;
-  if (targetSession.archived) return null;
-  if (targetSession.branch_id === sourceSession.branch_id) return null;
-
-  return {
-    ...targetSession,
-    branch_id: sourceSession.branch_id,
-    genealogy: {
-      ...(targetSession.genealogy ?? {}),
-      parent_session_id: sourceSession.session_id,
-    },
-    remote_surrogate: {
-      relationship,
-      source_session_id: sourceSession.session_id,
-      source_branch_id: sourceSession.branch_id,
-      target_branch_id: targetSession.branch_id,
-    },
-  };
-}
-
-function findSessionInBranchBuckets(
-  sessionsByBranchId: Map<string, Session[]>,
-  sessionId: string
-): Session | undefined {
-  for (const bucket of sessionsByBranchId.values()) {
-    const session = bucket.find((candidate) => candidate.session_id === sessionId);
-    if (session && !session.remote_surrogate) return session;
-  }
-  return undefined;
-}
-
 export function useAgorData(
   client: AgorClient | null,
   options?: { enabled?: boolean; directSessionId?: string | null }
 ): UseAgorDataResult {
   const enabled = options?.enabled ?? true;
   const directSessionId = options?.directSessionId ?? null;
-  // Single state for all server-backed maps — reset is setMaps(EMPTY_MAPS), one call, can't miss a field.
-  const [maps, setMaps] = useState<DataMaps>(EMPTY_MAPS);
 
-  // Per-field setter factory. Returns a setter with the same functional-update
-  // API as `useState`, with a no-op short-circuit: when the inner update
-  // returns the same reference for its slice, we preserve the outer `maps`
-  // reference too. Without this, `{ ...m, key: same }` would always allocate
-  // a fresh `maps` and force every `useAppLiveData()` / `useAppRepoData()`
-  // consumer to re-render on socket events the handler decided to discard.
-  const setMapSlice =
-    <K extends keyof DataMaps>(key: K) =>
-    (value: DataMaps[K] | ((prev: DataMaps[K]) => DataMaps[K])) =>
-      setMaps((prev) => {
-        const next =
-          typeof value === 'function'
-            ? (value as (p: DataMaps[K]) => DataMaps[K])(prev[key])
-            : value;
-        return Object.is(next, prev[key]) ? prev : { ...prev, [key]: next };
-      });
-  const setSessionById = setMapSlice('sessionById');
-  const setSessionsByBranch = setMapSlice('sessionsByBranch');
-  const setBoardById = setMapSlice('boardById');
-  const setCommentById = setMapSlice('commentById');
-  const setCardById = setMapSlice('cardById');
-  const setCardTypeById = setMapSlice('cardTypeById');
-  const setRepoById = setMapSlice('repoById');
-  const setBranchById = setMapSlice('branchById');
-  const setUserById = setMapSlice('userById');
-  const setMcpServerById = setMapSlice('mcpServerById');
-  const setGatewayChannelById = setMapSlice('gatewayChannelById');
-  const setArtifactById = setMapSlice('artifactById');
-  const setSessionMcpServerIds = setMapSlice('sessionMcpServerIds');
-  const setUserAuthenticatedMcpServerIds = setMapSlice('userAuthenticatedMcpServerIds');
-  const [loading, setLoading] = useState(true);
-  const [loadingStage, setLoadingStage] = useState<InitialLoadingStage>('idle');
-  const [error, setError] = useState<string | null>(null);
-  // Per-item counts captured at fetch-resolution time. Presence in this
-  // record means the item is "done"; the value is the size of the fetched
-  // list. Done flag and count flip atomically so a row never shows a green
-  // ✓ next to a stale 0 (the byId maps below are only populated after the
-  // full Promise.all resolves).
-  const [itemCounts, setItemCounts] = useState<Partial<Record<InitialLoadItemKey, number>>>({});
+  // Reset the shared singleton store once per hook (re)mount, synchronously
+  // BEFORE the first `useStore` read below. This mirrors the old per-mount
+  // `useState(EMPTY_MAPS)` / `useState(true)` semantics: the store is a module
+  // singleton (so a remount — and each test's `renderHook` — would otherwise
+  // inherit stale state), and `useAgorData` is its sole owner (mounted once in
+  // App.tsx). The `useState` initializer runs exactly once per instance.
+  //
+  // `resetHydrationRevisions()` zeroes the per-collection live-write baseline
+  // (fresh-`useRef` semantics); `cancelAllHydrations()` supersedes any straggler
+  // loop from a prior mount of the singleton (generations stay monotonic so a
+  // stale loop can never collide with this instance's fresh generation).
+  useState(() => {
+    agorStore.getState().reset();
+    resetHydrationRevisions();
+    cancelAllHydrations();
+    return null;
+  });
+
+  // Full store state, re-rendering this owner on every committed store change —
+  // exactly as the old single `useState<DataMaps>` + meta `useState`s did.
+  const storeState = useStore(agorStore);
 
   // Track if we've done initial fetch. The initial fetch happens once on mount;
   // socket reconnects after that re-trigger fetchData() to recover any events
@@ -650,57 +267,6 @@ export function useAgorData(
   // physical reconnect or page refresh. We use a ref rather than state since
   // we only consume it in event handlers, never in render.
   const lastSilentFetchFailedRef = useRef(false);
-
-  // Per-collection live-write revision counters — the core of the
-  // skip-apply-on-race background hydration. EVERY realtime handler that
-  // mutates one of these collection Maps bumps the matching counter
-  // (created / patched / removed, INCLUDING cascade removes such as branch
-  // eviction dropping its sessions, the deep-link-healing effect, and
-  // reconnect-driven writes). A background hydration snapshots the counters for
-  // the collections it replaces, fetches the full set, then applies the
-  // snapshot WHOLESALE only if those counters are unchanged when the fetch
-  // resolves — proving no live write raced. If any raced, the snapshot is
-  // discarded and refetched (never overlaid/reconciled). This makes a wholesale
-  // apply provably unable to clobber a live write OR resurrect a removed entity:
-  // a remove would have bumped the counter, so no apply happens. A ref (not
-  // state): only touched by async fetch code + event handlers, never in render.
-  const liveRevisionsRef = useRef<Record<HydratedCollection, number>>({
-    sessions: 0,
-    branches: 0,
-    boardObjects: 0,
-    cards: 0,
-    comments: 0,
-    mcpServers: 0,
-    sessionMcp: 0,
-    gatewayChannels: 0,
-    artifacts: 0,
-    oauth: 0,
-  });
-  // Stable bump helper (identity never changes — only mutates the ref) so it's
-  // safe to reference from the subscribe effect's handlers without churning deps.
-  const bumpRevision = useCallback((collection: HydratedCollection) => {
-    liveRevisionsRef.current[collection] += 1;
-  }, []);
-
-  // Per-collection hydration generation tokens. Each `runHydration` call bumps
-  // the generation for the collection(s) it owns and captures it; its retry loop
-  // stops (without applying a snapshot or scheduling another timer) the moment a
-  // newer hydration supersedes it (a reconnect-triggered refetch), the component
-  // unmounts, or a logout reset fires — all of which bump these counters. This
-  // is CANCELLATION, not race reconciliation: clobber-safety still comes entirely
-  // from the quiet-window check against `liveRevisionsRef`.
-  const hydrationGenerationRef = useRef<Record<HydratedCollection, number>>({
-    sessions: 0,
-    branches: 0,
-    boardObjects: 0,
-    cards: 0,
-    comments: 0,
-    mcpServers: 0,
-    sessionMcp: 0,
-    gatewayChannels: 0,
-    artifacts: 0,
-    oauth: 0,
-  });
 
   // Fetch all data
   //
@@ -727,11 +293,11 @@ export function useAgorData(
 
       try {
         if (!silent) {
-          setLoading(true);
-          setLoadingStage('fetching');
+          agorStore.getState().setLoading(true);
+          agorStore.getState().setLoadingStage('fetching');
           debugTimer?.markStage('fetching');
-          setError(null);
-          setItemCounts({});
+          agorStore.getState().setError(null);
+          agorStore.getState().setItemCounts({});
         }
 
         // Marks a tracked item complete (and captures its count from the
@@ -743,78 +309,10 @@ export function useAgorData(
         ): Promise<T> => {
           const timedPromise = debugTimer?.track(key, p) ?? p;
           return timedPromise.then((r) => {
-            if (!silent) setItemCounts((prev) => ({ ...prev, [key]: r.length }));
+            if (!silent)
+              agorStore.getState().setItemCounts((prev) => ({ ...prev, [key]: r.length }));
             return r;
           });
-        };
-
-        // Run a BACKGROUND (non-gated) hydration with skip-apply-on-race. The
-        // fetched full-set snapshot is applied WHOLESALE only if no live write
-        // to any of `collections` raced the fetch — proven by snapshotting each
-        // collection's revision counter before the fetch and re-checking after.
-        // If a write raced, the (potentially stale) snapshot is DISCARDED and
-        // refetched from a fresh baseline; we NEVER overlay or reconcile a racy
-        // snapshot. It retries until it lands a quiet window — a few immediate
-        // retries then capped exponential backoff — and never gives up (skipping
-        // forever could leave Home empty/incomplete indefinitely; live events
-        // only deliver changes, not backfill). The loop is cancelled — not
-        // abandoned — on supersession (reconnect), unmount, or logout reset.
-        const runHydration = async <T>(
-          label: string,
-          collections: readonly HydratedCollection[],
-          fetchFn: () => Promise<T>,
-          apply: (result: T) => void
-        ): Promise<void> => {
-          // Supersede any older loop for these collections and capture our
-          // generation token. The loop bails the instant a newer hydration
-          // (reconnect), an unmount, or a logout reset bumps the generation — so
-          // it never applies a stale snapshot or schedules another timer after
-          // it's been cancelled.
-          const myGeneration = collections.map((c) => (hydrationGenerationRef.current[c] += 1));
-          const isCurrent = () =>
-            collections.every((c, i) => hydrationGenerationRef.current[c] === myGeneration[i]);
-          // Delay PRECEDING attempt N: the first HYDRATION_IMMEDIATE_RETRIES
-          // attempts fire back-to-back (delay 0) so a single transient race
-          // converges instantly; after that, capped exponential backoff lets a
-          // sustained write burst settle.
-          const delayBeforeAttempt = (attempt: number) =>
-            attempt < HYDRATION_IMMEDIATE_RETRIES
-              ? 0
-              : Math.min(
-                  HYDRATION_BACKOFF_BASE_MS * 2 ** (attempt - HYDRATION_IMMEDIATE_RETRIES),
-                  HYDRATION_BACKOFF_CAP_MS
-                );
-
-          // Retry until a quiet-window apply SUCCEEDS (or the loop is cancelled).
-          // We never force-apply a racy snapshot — we just keep re-snapshotting
-          // and re-fetching until no live write races a fetch.
-          for (let attempt = 0; ; attempt++) {
-            const delayMs = delayBeforeAttempt(attempt);
-            if (delayMs > 0) {
-              await new Promise<void>((resolve) => setTimeout(resolve, delayMs));
-              if (!isCurrent()) return; // superseded while waiting
-            }
-            const before = collections.map((c) => liveRevisionsRef.current[c]);
-            let result: T;
-            try {
-              result = await fetchFn();
-            } catch (err) {
-              console.warn(`[useAgorData] background ${label} fetch failed:`, err);
-              if (!isCurrent()) return; // superseded while fetching
-              // A failed fetch leaves the collection un-hydrated; retrying (with
-              // backoff) is exactly what keeps Home from staying empty forever.
-              continue;
-            }
-            if (!isCurrent()) return; // superseded while fetching
-            const raced = collections.some((c, i) => liveRevisionsRef.current[c] !== before[i]);
-            if (!raced) {
-              apply(result);
-              return;
-            }
-            // A live write to one of these collections raced the fetch — discard
-            // this snapshot and retry from a fresh revision baseline (the next
-            // iteration's delay precedes its fetch).
-          }
         };
 
         // ── Background (non-gated) fetches ──────────────────────────────
@@ -826,16 +324,18 @@ export function useAgorData(
         // snapshot when no live write to that collection raced (else it refetches
         // a fresh one). We deliberately do NOT `track()` them — they're absent
         // from INITIAL_LOAD_ITEMS, so the loading checklist / `initialLoadComplete`
-        // gate ignores them. We apply through the stable `setMaps` (not the
-        // per-render setMapSlice setters, which would destabilize this
-        // useCallback's deps and re-fire the subscribe effect).
+        // gate ignores them. We apply through the store's `applyMaps` (not the
+        // per-entity setters), keeping fetchData's deps stable so the subscribe
+        // effect doesn't re-fire.
         void runHydration(
           'mcp-servers',
           ['mcpServers'],
           () =>
             client.service('mcp-servers').findAll({ query: { $limit: PAGINATION.DEFAULT_LIMIT } }),
           (list) =>
-            setMaps((prev) => ({ ...prev, mcpServerById: buildById(list, 'mcp_server_id') }))
+            agorStore
+              .getState()
+              .applyMaps((prev) => ({ ...prev, mcpServerById: buildById(list, 'mcp_server_id') }))
         );
         void runHydration(
           'session-mcp-servers',
@@ -844,7 +344,10 @@ export function useAgorData(
             client
               .service('session-mcp-servers')
               .findAll({ query: { $limit: PAGINATION.DEFAULT_LIMIT } }),
-          (list) => setMaps((prev) => ({ ...prev, sessionMcpServerIds: buildSessionMcpMap(list) }))
+          (list) =>
+            agorStore
+              .getState()
+              .applyMaps((prev) => ({ ...prev, sessionMcpServerIds: buildSessionMcpMap(list) }))
         );
         void runHydration(
           'gateway-channels',
@@ -853,7 +356,10 @@ export function useAgorData(
             client
               .service('gateway-channels')
               .findAll({ query: { $limit: PAGINATION.DEFAULT_LIMIT } }),
-          (list) => setMaps((prev) => ({ ...prev, gatewayChannelById: buildById(list, 'id') }))
+          (list) =>
+            agorStore
+              .getState()
+              .applyMaps((prev) => ({ ...prev, gatewayChannelById: buildById(list, 'id') }))
         );
         void runHydration(
           'artifacts',
@@ -885,7 +391,10 @@ export function useAgorData(
                 ],
               },
             }),
-          (list) => setMaps((prev) => ({ ...prev, artifactById: buildById(list, 'artifact_id') }))
+          (list) =>
+            agorStore
+              .getState()
+              .applyMaps((prev) => ({ ...prev, artifactById: buildById(list, 'artifact_id') }))
         );
         void runHydration(
           'oauth-status',
@@ -894,7 +403,9 @@ export function useAgorData(
           (res) => {
             const ids =
               (res as { authenticated_server_ids?: string[] })?.authenticated_server_ids ?? [];
-            setMaps((prev) => ({ ...prev, userAuthenticatedMcpServerIds: new Set(ids) }));
+            agorStore
+              .getState()
+              .applyMaps((prev) => ({ ...prev, userAuthenticatedMcpServerIds: new Set(ids) }));
           }
         );
 
@@ -1125,7 +636,7 @@ export function useAgorData(
         debugTimer?.endFetchPhase();
 
         if (!silent) {
-          setLoadingStage('indexing');
+          agorStore.getState().setLoadingStage('indexing');
           debugTimer?.markStage('indexing');
           debugTimer?.startIndexing();
           // Give the browser one paint opportunity so large instances can
@@ -1195,7 +706,7 @@ export function useAgorData(
         // sessionMcpServerIds / userAuthenticatedMcpServerIds — survive even if
         // their fire-and-forget fetches resolved before this gate did. Those
         // slices are owned by their background setters + realtime handlers.
-        setMaps((prev) => ({
+        agorStore.getState().applyMaps((prev) => ({
           ...prev,
           sessionById: sessionsById,
           sessionsByBranch: sessionsByBranchId,
@@ -1219,9 +730,7 @@ export function useAgorData(
         // snapshot (resurrecting data that changed while we were disconnected).
         // The background hydrations kicked off below re-snapshot AFTER this bump,
         // so they're unaffected.
-        for (const c of ['sessions', 'branches', 'boardObjects', 'cards', 'comments'] as const) {
-          liveRevisionsRef.current[c] += 1;
-        }
+        bumpFirstPaintMergeRevisions();
         debugTimer?.endIndexing();
         debugFinishStatus = 'success';
 
@@ -1267,7 +776,7 @@ export function useAgorData(
                 },
               }),
             (allSessions) =>
-              setMaps((prev) => {
+              agorStore.getState().applyMaps((prev) => {
                 // The hydration fetches active sessions only. Deep-link-healed
                 // archived sessions (added to `sessionById` so a direct /s/<id>
                 // archived link can open the drawer) are OUT of that query's
@@ -1296,7 +805,9 @@ export function useAgorData(
               // are active-only (the snapshot query is archived:false and the
               // handlers never keep an archived branch), so a wholesale replace
               // is complete.
-              setMaps((prev) => ({ ...prev, branchById: buildById(allBranches, 'branch_id') }))
+              agorStore
+                .getState()
+                .applyMaps((prev) => ({ ...prev, branchById: buildById(allBranches, 'branch_id') }))
           );
         }
 
@@ -1317,7 +828,7 @@ export function useAgorData(
                 .service('board-objects')
                 .findAll({ query: { $limit: PAGINATION.DEFAULT_LIMIT } }),
             (allBoardObjects) =>
-              setMaps((prev) => {
+              agorStore.getState().applyMaps((prev) => {
                 const base = buildBoardObjectMaps(allBoardObjects);
                 return {
                   ...prev,
@@ -1332,7 +843,10 @@ export function useAgorData(
             'cards',
             ['cards'],
             () => client.service('cards').findAll({ query: { $limit: PAGINATION.DEFAULT_LIMIT } }),
-            (allCards) => setMaps((prev) => ({ ...prev, cardById: buildById(allCards, 'card_id') }))
+            (allCards) =>
+              agorStore
+                .getState()
+                .applyMaps((prev) => ({ ...prev, cardById: buildById(allCards, 'card_id') }))
           );
           void runHydration(
             'board-comments',
@@ -1342,7 +856,10 @@ export function useAgorData(
                 .service('board-comments')
                 .findAll({ query: { $limit: PAGINATION.DEFAULT_LIMIT } }),
             (allComments) =>
-              setMaps((prev) => ({ ...prev, commentById: buildById(allComments, 'comment_id') }))
+              agorStore.getState().applyMaps((prev) => ({
+                ...prev,
+                commentById: buildById(allComments, 'comment_id'),
+              }))
           );
         }
 
@@ -1362,12 +879,14 @@ export function useAgorData(
         } else {
           debugFinishStatus = 'error';
           debugFinishError = err;
-          setError(err instanceof Error ? err.message : 'Failed to fetch data');
+          agorStore
+            .getState()
+            .setError(err instanceof Error ? err.message : 'Failed to fetch data');
         }
       } finally {
         if (!silent) {
-          setLoading(false);
-          setLoadingStage('idle');
+          agorStore.getState().setLoading(false);
+          agorStore.getState().setLoadingStage('idle');
           debugTimer?.markStage('idle');
           if (debugFinishStatus) {
             debugTimer?.finish(debugFinishStatus, debugFinishError);
@@ -1388,44 +907,35 @@ export function useAgorData(
   // board (and downstream, the URL) — see the comment on the useAgorData
   // call in App.tsx for the full failure chain.
   //
-  // EMPTY_MAPS covers every field — adding a new map to DataMaps automatically
-  // includes it here without any extra code.
+  // `resetMaps()` clears every data map (EMPTY_MAPS covers every field) while
+  // leaving the meta fields alone — matching the old `setMaps(EMPTY_MAPS)`.
+  // `cancelAndFailAllHydrations()` cancels every in-flight hydration loop (bump
+  // generations) AND fails any quiet check it might still reach (bump revisions)
+  // so an unresolved hydration can't repopulate the Maps AFTER logout (post-logout
+  // data leak). Bumping the generation is the real stop — without it, a revision
+  // bump alone would only make the loop discard-and-RE-FETCH from the stale client
+  // and eventually apply into freshly-cleared Maps.
   useEffect(() => {
     if (client) return;
-    // Cancel every in-flight hydration loop (bump generations) AND fail any
-    // quiet check it might still reach (bump revisions) so an unresolved
-    // hydration can't repopulate the Maps AFTER logout (post-logout data leak).
-    // Bumping the generation is the real stop — without it, a revision bump alone
-    // would only make the loop discard-and-RE-FETCH from the stale client and
-    // eventually apply into freshly-cleared Maps. Same lynchpin as the
-    // per-mutation revision bumps.
-    for (const c of Object.keys(liveRevisionsRef.current) as HydratedCollection[]) {
-      hydrationGenerationRef.current[c] += 1;
-      liveRevisionsRef.current[c] += 1;
-    }
-    setMaps(EMPTY_MAPS);
+    cancelAndFailAllHydrations();
+    agorStore.getState().resetMaps();
     setHasInitiallyFetched(false);
   }, [client]);
 
   // On unmount, supersede every in-flight per-collection hydration loop so it
   // stops retrying and never applies a snapshot (or schedules another timer)
   // after teardown. Generation bump = cancellation; see `runHydration`.
-  useEffect(() => {
-    const generations = hydrationGenerationRef.current;
-    return () => {
-      for (const c of Object.keys(generations) as HydratedCollection[]) {
-        generations[c] += 1;
-      }
-    };
-  }, []);
+  useEffect(() => () => cancelAllHydrations(), []);
 
   // If the user navigates to /s/<id>/ after the initial active-session fetch,
   // load that one session by ID as well. This keeps direct links to archived
   // sessions openable without changing the default list query.
   useEffect(() => {
     if (!client || !enabled || !hasInitiallyFetched || !directSessionId) return;
-    if (maps.sessionById.has(directSessionId)) return;
-    if (hasIdMatchingPrefix(directSessionId, maps.sessionById.values(), (s) => s.session_id)) {
+    if (storeState.sessionById.has(directSessionId)) return;
+    if (
+      hasIdMatchingPrefix(directSessionId, storeState.sessionById.values(), (s) => s.session_id)
+    ) {
       return;
     }
 
@@ -1439,14 +949,14 @@ export function useAgorData(
         // hydration in flight discards its (session-missing) snapshot rather
         // than clobbering this deep-link heal.
         bumpRevision('sessions');
-        setSessionById((prev) => {
+        agorStore.getState().setMap('sessionById', (prev) => {
           if (prev.has(directSession.session_id)) return prev;
           const next = new Map(prev);
           next.set(directSession.session_id, directSession);
           return next;
         });
         if (!directSession.archived) {
-          setSessionsByBranch((prev) => {
+          agorStore.getState().setMap('sessionsByBranch', (prev) => {
             const branchSessions = prev.get(directSession.branch_id) || [];
             if (branchSessions.some((s) => s.session_id === directSession.session_id)) return prev;
             const next = new Map(prev);
@@ -1458,7 +968,7 @@ export function useAgorData(
         if (
           !directSession.archived &&
           directSession.branch_id &&
-          !maps.branchById.has(directSession.branch_id)
+          !storeState.branchById.has(directSession.branch_id)
         ) {
           try {
             const directBranch = (await client
@@ -1466,7 +976,7 @@ export function useAgorData(
               .get(directSession.branch_id)) as Branch;
             if (cancelled) return;
             bumpRevision('branches');
-            setBranchById((prev) => {
+            agorStore.getState().setMap('branchById', (prev) => {
               if (directBranch.archived) return prev;
               if (prev.has(directBranch.branch_id)) return prev;
               const next = new Map(prev);
@@ -1487,577 +997,106 @@ export function useAgorData(
       cancelled = true;
     };
   }, [
-    bumpRevision,
     client,
     directSessionId,
     enabled,
     hasInitiallyFetched,
-    maps.branchById,
-    maps.sessionById,
-    setBranchById,
-    setSessionById,
-    setSessionsByBranch,
+    storeState.branchById,
+    storeState.sessionById,
   ]);
 
   // Subscribe to real-time updates
-  // biome-ignore lint/correctness/useExhaustiveDependencies: setter helpers only close over stable setMaps; listing them would add noise without preventing stale closures
+  //
+  // Every socket event is wired straight to the matching store action in
+  // `agorRealtimeActions` (module singletons → stable references, so cleanup
+  // `removeListener` matches). The store action does the same
+  // `replaceIfChanged` / cascade / index-rebuild + per-collection `bumpRevision`
+  // the hook used to do inline. OAuth + agor-query handlers stay local: they need
+  // `client` (async refetch) or are pure window side-effects.
   useEffect(() => {
     if (!client || !enabled) {
       // No client or disabled = not ready for data fetch, set loading to false
-      setLoading(false);
-      setLoadingStage('idle');
+      agorStore.getState().setLoading(false);
+      agorStore.getState().setLoadingStage('idle');
       return;
     }
 
     // Subscribe to session events
     const sessionsService = client.service('sessions');
-    const handleSessionCreated = (session: Session) => {
-      // Bump the sessions revision so an in-flight sessions hydration discards
-      // its snapshot and refetches instead of clobbering this write.
-      bumpRevision('sessions');
-      if (session.archived) return;
-
-      // Update sessionById - only create new Map if session doesn't exist
-      setSessionById((prev) => {
-        if (prev.has(session.session_id)) return prev; // Already exists, shouldn't happen
-        const next = new Map(prev);
-        next.set(session.session_id, session);
-        return next;
-      });
-
-      // Update sessionsByBranch - only create new Map when adding new session
-      setSessionsByBranch((prev) => {
-        const branchSessions = prev.get(session.branch_id) || [];
-        // Check if session already exists in this branch (duplicate event)
-        if (branchSessions.some((s) => s.session_id === session.session_id)) return prev;
-
-        const next = new Map(prev);
-        next.set(session.branch_id, [...branchSessions, session]);
-        return next;
-      });
-    };
-    const handleSessionPatched = (session: Session) => {
-      // Patch (incl. archive, which removes the session from the active maps)
-      // counts as a live write — bump so an in-flight sessions hydration can't
-      // clobber it or resurrect an archive with a pre-archive snapshot.
-      bumpRevision('sessions');
-      const isArchived = session.archived === true;
-      // Track old branch_id for migration detection
-      let oldBranchId: string | null = null;
-
-      // Update sessionById - add/update active sessions, remove archived sessions
-      setSessionById((prev) => {
-        const existing = prev.get(session.session_id);
-
-        // Capture old branch_id before updating
-        oldBranchId = existing?.branch_id || null;
-
-        if (isArchived) {
-          if (!existing) return prev;
-          const next = new Map(prev);
-          next.delete(session.session_id);
-          return next;
-        }
-
-        const mergedSession = preserveSessionRelationshipFields(session, existing);
-
-        // Bail out on no-op patches. Feathers always emits a fresh object so
-        // `existing === session` never holds, but the daemon does emit
-        // idempotent patches (e.g. callback bookkeeping that lands at the same
-        // status). Shallow-equal misses nested fields the daemon reserializes
-        // — that's a safe false negative.
-        if (existing && shallowEqualEntity(existing, mergedSession)) return prev;
-
-        const next = new Map(prev);
-        next.set(session.session_id, mergedSession);
-        return next;
-      });
-
-      // Update sessionsByBranch - keep active sessions only
-      setSessionsByBranch((prev) => {
-        let changed = false;
-        const next = new Map(prev);
-        const newBranchId = session.branch_id;
-
-        const removeFromBranch = (branchId: string) => {
-          const bucket = next.get(branchId) || [];
-          const filtered = bucket.filter((s) => s.session_id !== session.session_id);
-          if (filtered.length !== bucket.length) {
-            changed = true;
-            if (filtered.length > 0) {
-              next.set(branchId, filtered);
-            } else {
-              next.delete(branchId);
-            }
-          }
-        };
-
-        if (isArchived) {
-          for (const [branchId, bucket] of next) {
-            if (bucket.some((item) => item.session_id === session.session_id)) {
-              removeFromBranch(branchId);
-            }
-          }
-          return changed ? next : prev;
-        }
-
-        // Session moved between branches - remove from old bucket first
-        const branchMigrated = oldBranchId && oldBranchId !== newBranchId;
-        if (branchMigrated) {
-          removeFromBranch(oldBranchId!);
-        }
-
-        const branchSessions = next.get(newBranchId) || [];
-        const index = branchSessions.findIndex((s) => s.session_id === session.session_id);
-        let sourceSessionForRemoteProjection = session;
-
-        if (index === -1) {
-          next.set(newBranchId, [...branchSessions, session]);
-        } else {
-          const mergedSession = preserveSessionRelationshipFields(session, branchSessions[index]);
-          sourceSessionForRemoteProjection = mergedSession;
-
-          // Bail out when the session is content-equal to what we already hold.
-          // Mirrors the sessionById bailout above so an idempotent patch doesn't
-          // produce a fresh branch-bucket array (which would invalidate
-          // `data.sessions === n.sessions` in BranchNode's custom areEqual and
-          // re-render every BranchCard on the affected branch).
-          if (
-            branchSessions[index] === mergedSession ||
-            shallowEqualEntity(branchSessions[index], mergedSession)
-          ) {
-            return changed ? next : prev;
-          }
-
-          const updatedSessions = [...branchSessions];
-          updatedSessions[index] = mergedSession;
-          next.set(newBranchId, updatedSessions);
-
-          // Also update any remote/surrogate projections of this session that
-          // live in source-branch buckets. Preserve their local tree placement
-          // while refreshing status/callback_config/etc. from the canonical row.
-          for (const [branchId, bucket] of next) {
-            if (branchId === newBranchId) continue;
-
-            let bucketChanged = false;
-            const refreshedBucket = bucket.map((item) => {
-              if (item.session_id !== session.session_id) return item;
-              bucketChanged = true;
-              return {
-                ...preserveSessionRelationshipFields(session, item),
-                branch_id: item.branch_id,
-                genealogy: item.genealogy,
-                remote_surrogate: item.remote_surrogate,
-              };
-            });
-
-            if (bucketChanged) {
-              next.set(branchId, refreshedBucket);
-            }
-          }
-        }
-
-        // Remote relationships are created after the canonical target session
-        // row. The daemon then emits a patched source session with
-        // remote_relationships.as_source populated. Project that single source
-        // row into muted remote-surrogate children now, instead of doing any
-        // expensive relationship work during render.
-        for (const relationship of sourceSessionForRemoteProjection.remote_relationships
-          ?.as_source ?? []) {
-          if (relationship.relationship_type !== 'remote_create') continue;
-
-          const targetSession = findSessionInBranchBuckets(next, relationship.target_session_id);
-          if (!targetSession) continue;
-
-          const sourceBranchSessions = next.get(sourceSessionForRemoteProjection.branch_id) ?? [];
-          if (
-            sourceBranchSessions.some(
-              (candidate) => candidate.session_id === targetSession.session_id
-            )
-          ) {
-            continue;
-          }
-
-          const remoteSurrogate = createRemoteSurrogateSession(
-            sourceSessionForRemoteProjection,
-            targetSession,
-            relationship
-          );
-          if (!remoteSurrogate) continue;
-
-          next.set(sourceSessionForRemoteProjection.branch_id, [
-            ...sourceBranchSessions,
-            remoteSurrogate,
-          ]);
-        }
-
-        return next;
-      });
-    };
-    const handleSessionRemoved = (session: Session) => {
-      bumpRevision('sessions');
-      // Update sessionById — bail out when the id isn't tracked so the
-      // wrapper short-circuit prevents the spurious `maps` update.
-      setSessionById((prev) => {
-        if (!prev.has(session.session_id)) return prev;
-        const next = new Map(prev);
-        next.delete(session.session_id);
-        return next;
-      });
-
-      // Update sessionsByBranch — same bail when the session isn't in the
-      // branch's bucket.
-      setSessionsByBranch((prev) => {
-        const branchSessions = prev.get(session.branch_id);
-        if (!branchSessions?.some((s) => s.session_id === session.session_id)) {
-          return prev;
-        }
-        const next = new Map(prev);
-        const filtered = branchSessions.filter((s) => s.session_id !== session.session_id);
-        if (filtered.length > 0) {
-          next.set(session.branch_id, filtered);
-        } else {
-          // Clean up empty arrays
-          next.delete(session.branch_id);
-        }
-        return next;
-      });
-    };
-
-    sessionsService.on('created', handleSessionCreated);
-    sessionsService.on('patched', handleSessionPatched);
-    sessionsService.on('updated', handleSessionPatched);
-    sessionsService.on('removed', handleSessionRemoved);
+    sessionsService.on('created', realtime.sessionCreated);
+    sessionsService.on('patched', realtime.sessionPatched);
+    sessionsService.on('updated', realtime.sessionPatched);
+    sessionsService.on('removed', realtime.sessionRemoved);
 
     // Subscribe to board events
     const boardsService = client.service('boards');
-    const handleBoardCreated = (board: Board) => {
-      setBoardById((prev) => {
-        if (prev.has(board.board_id)) return prev; // Already exists, shouldn't happen
-        const next = new Map(prev);
-        next.set(board.board_id, board);
-        return next;
-      });
-    };
-    const handleBoardPatched = (board: Board) => {
-      setBoardById((prev) => replaceIfChanged(prev, board.board_id, board));
-    };
-    const handleBoardRemoved = (board: Board) => {
-      setBoardById((prev) => {
-        if (!prev.has(board.board_id)) return prev; // Doesn't exist, nothing to remove
-        const next = new Map(prev);
-        next.delete(board.board_id);
-        return next;
-      });
-    };
-
-    boardsService.on('created', handleBoardCreated);
-    boardsService.on('patched', handleBoardPatched);
-    boardsService.on('updated', handleBoardPatched);
-    boardsService.on('removed', handleBoardRemoved);
+    boardsService.on('created', realtime.boardCreated);
+    boardsService.on('patched', realtime.boardPatched);
+    boardsService.on('updated', realtime.boardPatched);
+    boardsService.on('removed', realtime.boardRemoved);
 
     // Subscribe to board object events
     const boardObjectsService = client.service('board-objects');
-    const handleBoardObjectCreated = (boardObject: BoardEntityObject) => {
-      bumpRevision('boardObjects');
-      setMaps((prev) => upsertBoardObjectInMaps(prev, boardObject, 'create'));
-    };
-    const handleBoardObjectPatched = (boardObject: BoardEntityObject) => {
-      bumpRevision('boardObjects');
-      setMaps((prev) => upsertBoardObjectInMaps(prev, boardObject, 'patch'));
-    };
-    const handleBoardObjectRemoved = (boardObject: BoardEntityObject) => {
-      bumpRevision('boardObjects');
-      setMaps((prev) => removeBoardObjectFromMaps(prev, boardObject));
-    };
-
-    boardObjectsService.on('created', handleBoardObjectCreated);
-    boardObjectsService.on('patched', handleBoardObjectPatched);
-    boardObjectsService.on('updated', handleBoardObjectPatched);
-    boardObjectsService.on('removed', handleBoardObjectRemoved);
+    boardObjectsService.on('created', realtime.boardObjectCreated);
+    boardObjectsService.on('patched', realtime.boardObjectPatched);
+    boardObjectsService.on('updated', realtime.boardObjectPatched);
+    boardObjectsService.on('removed', realtime.boardObjectRemoved);
 
     // Subscribe to repo events
     const reposService = client.service('repos');
-    const handleRepoCreated = (repo: Repo) => {
-      setRepoById((prev) => {
-        if (prev.has(repo.repo_id)) return prev; // Already exists, shouldn't happen
-        const next = new Map(prev);
-        next.set(repo.repo_id, repo);
-        return next;
-      });
-    };
-    const handleRepoPatched = (repo: Repo) => {
-      setRepoById((prev) => replaceIfChanged(prev, repo.repo_id, repo));
-    };
-    const handleRepoRemoved = (repo: Repo) => {
-      setRepoById((prev) => {
-        if (!prev.has(repo.repo_id)) return prev; // Doesn't exist, nothing to remove
-        const next = new Map(prev);
-        next.delete(repo.repo_id);
-        return next;
-      });
-    };
-
-    reposService.on('created', handleRepoCreated);
-    reposService.on('patched', handleRepoPatched);
-    reposService.on('updated', handleRepoPatched);
-    reposService.on('removed', handleRepoRemoved);
+    reposService.on('created', realtime.repoCreated);
+    reposService.on('patched', realtime.repoPatched);
+    reposService.on('updated', realtime.repoPatched);
+    reposService.on('removed', realtime.repoRemoved);
 
     // Subscribe to branch events
     const branchesService = client.service('branches');
-    const handleBranchCreated = (branch: Branch) => {
-      // Bump the branches revision so an in-flight branches hydration can't
-      // clobber this write (mirrors the session handlers).
-      bumpRevision('branches');
-      if (branch.archived) return;
-
-      setBranchById((prev) => {
-        if (prev.has(branch.branch_id)) return prev; // Already exists, shouldn't happen
-        const next = new Map(prev);
-        next.set(branch.branch_id, branch);
-        return next;
-      });
-    };
-    // Drop a branch from `branchById` and prune every session that lived on
-    // it from `sessionById` / `sessionsByBranch`. Shared between the
-    // `archived: true` patch path and the hard-delete `removed` path —
-    // either way we never want an orphan session card to linger.
-    const evictBranchAndSessions = (branchId: string) => {
-      // This cascade mutates BOTH the branches map (caller already bumped) and
-      // the sessions maps, so bump the sessions revision too — otherwise a
-      // sessions hydration in flight could resurrect the evicted sessions with a
-      // pre-eviction snapshot.
-      bumpRevision('sessions');
-      setBranchById((prev) => {
-        if (!prev.has(branchId)) return prev;
-        const next = new Map(prev);
-        next.delete(branchId);
-        return next;
-      });
-      setSessionsByBranch((prev) => {
-        if (!prev.has(branchId)) return prev;
-        const next = new Map(prev);
-        next.delete(branchId);
-        return next;
-      });
-      setSessionById((prev) => {
-        let changed = false;
-        const next = new Map(prev);
-        for (const [sessionId, session] of prev.entries()) {
-          if (session.branch_id === branchId) {
-            next.delete(sessionId);
-            changed = true;
-          }
-        }
-        return changed ? next : prev;
-      });
-    };
-
-    const handleBranchPatched = (branch: Branch) => {
-      bumpRevision('branches');
-      if (branch.archived) {
-        evictBranchAndSessions(branch.branch_id);
-        return;
-      }
-
-      setBranchById((prev) => replaceIfChanged(prev, branch.branch_id, branch));
-    };
-    const handleBranchRemoved = (branch: Branch) => {
-      bumpRevision('branches');
-      // Mirror the archive path: a hard delete should also evict any
-      // sessions we still track on that branch.
-      evictBranchAndSessions(branch.branch_id);
-    };
-
-    branchesService.on('created', handleBranchCreated);
-    branchesService.on('patched', handleBranchPatched);
-    branchesService.on('updated', handleBranchPatched);
-    branchesService.on('removed', handleBranchRemoved);
+    branchesService.on('created', realtime.branchCreated);
+    branchesService.on('patched', realtime.branchPatched);
+    branchesService.on('updated', realtime.branchPatched);
+    branchesService.on('removed', realtime.branchRemoved);
 
     // Subscribe to user events
     const usersService = client.service('users');
-    const handleUserCreated = (user: User) => {
-      setUserById((prev) => {
-        if (prev.has(user.user_id)) return prev; // Already exists, shouldn't happen
-        const next = new Map(prev);
-        next.set(user.user_id, user);
-        return next;
-      });
-    };
-    const handleUserPatched = (user: User) => {
-      setUserById((prev) => replaceIfChanged(prev, user.user_id, user));
-    };
-    const handleUserRemoved = (user: User) => {
-      setUserById((prev) => {
-        if (!prev.has(user.user_id)) return prev; // Doesn't exist, nothing to remove
-        const next = new Map(prev);
-        next.delete(user.user_id);
-        return next;
-      });
-    };
-
-    usersService.on('created', handleUserCreated);
-    usersService.on('patched', handleUserPatched);
-    usersService.on('updated', handleUserPatched);
-    usersService.on('removed', handleUserRemoved);
+    usersService.on('created', realtime.userCreated);
+    usersService.on('patched', realtime.userPatched);
+    usersService.on('updated', realtime.userPatched);
+    usersService.on('removed', realtime.userRemoved);
 
     // Subscribe to MCP server events
     const mcpServersService = client.service('mcp-servers');
-    const handleMCPServerCreated = (server: MCPServer) => {
-      bumpRevision('mcpServers');
-      setMcpServerById((prev) => {
-        if (prev.has(server.mcp_server_id)) return prev; // Already exists, shouldn't happen
-        const next = new Map(prev);
-        next.set(server.mcp_server_id, server);
-        return next;
-      });
-    };
-    const handleMCPServerPatched = (server: MCPServer) => {
-      bumpRevision('mcpServers');
-      setMcpServerById((prev) => replaceIfChanged(prev, server.mcp_server_id, server));
-    };
-    const handleMCPServerRemoved = (server: MCPServer) => {
-      bumpRevision('mcpServers');
-      setMcpServerById((prev) => {
-        if (!prev.has(server.mcp_server_id)) return prev; // Doesn't exist, nothing to remove
-        const next = new Map(prev);
-        next.delete(server.mcp_server_id);
-        return next;
-      });
-    };
-
-    mcpServersService.on('created', handleMCPServerCreated);
-    mcpServersService.on('patched', handleMCPServerPatched);
-    mcpServersService.on('updated', handleMCPServerPatched);
-    mcpServersService.on('removed', handleMCPServerRemoved);
+    mcpServersService.on('created', realtime.mcpServerCreated);
+    mcpServersService.on('patched', realtime.mcpServerPatched);
+    mcpServersService.on('updated', realtime.mcpServerPatched);
+    mcpServersService.on('removed', realtime.mcpServerRemoved);
 
     // Subscribe to gateway channel events
     const gatewayChannelsService = client.service('gateway-channels');
-    const handleGatewayChannelCreated = (channel: GatewayChannel) => {
-      bumpRevision('gatewayChannels');
-      setGatewayChannelById((prev) => {
-        if (prev.has(channel.id)) return prev;
-        const next = new Map(prev);
-        next.set(channel.id, channel);
-        return next;
-      });
-    };
-    const handleGatewayChannelPatched = (channel: GatewayChannel) => {
-      bumpRevision('gatewayChannels');
-      setGatewayChannelById((prev) => replaceIfChanged(prev, channel.id, channel));
-    };
-    const handleGatewayChannelRemoved = (channel: GatewayChannel) => {
-      bumpRevision('gatewayChannels');
-      setGatewayChannelById((prev) => {
-        if (!prev.has(channel.id)) return prev;
-        const next = new Map(prev);
-        next.delete(channel.id);
-        return next;
-      });
-    };
-
-    gatewayChannelsService.on('created', handleGatewayChannelCreated);
-    gatewayChannelsService.on('patched', handleGatewayChannelPatched);
-    gatewayChannelsService.on('updated', handleGatewayChannelPatched);
-    gatewayChannelsService.on('removed', handleGatewayChannelRemoved);
+    gatewayChannelsService.on('created', realtime.gatewayChannelCreated);
+    gatewayChannelsService.on('patched', realtime.gatewayChannelPatched);
+    gatewayChannelsService.on('updated', realtime.gatewayChannelPatched);
+    gatewayChannelsService.on('removed', realtime.gatewayChannelRemoved);
 
     // Subscribe to card events
     const cardsService = client.service('cards');
-    const handleCardCreated = (card: CardWithType) => {
-      bumpRevision('cards');
-      setCardById((prev) => {
-        if (prev.has(card.card_id)) return prev; // Duplicate event — bail.
-        const next = new Map(prev);
-        next.set(card.card_id, card);
-        return next;
-      });
-    };
-    const handleCardPatched = (card: CardWithType) => {
-      bumpRevision('cards');
-      setCardById((prev) => replaceIfChanged(prev, card.card_id, card));
-    };
-    const handleCardRemoved = (card: CardWithType) => {
-      bumpRevision('cards');
-      setCardById((prev) => {
-        if (!prev.has(card.card_id)) return prev;
-        const next = new Map(prev);
-        next.delete(card.card_id);
-        return next;
-      });
-    };
-
-    cardsService.on('created', handleCardCreated);
-    cardsService.on('patched', handleCardPatched);
-    cardsService.on('updated', handleCardPatched);
-    cardsService.on('removed', handleCardRemoved);
+    cardsService.on('created', realtime.cardCreated);
+    cardsService.on('patched', realtime.cardPatched);
+    cardsService.on('updated', realtime.cardPatched);
+    cardsService.on('removed', realtime.cardRemoved);
 
     // Subscribe to card type events
     const cardTypesService = client.service('card-types');
-    const handleCardTypeCreated = (cardType: CardType) => {
-      setCardTypeById((prev) => {
-        if (prev.has(cardType.card_type_id)) return prev; // Duplicate event — bail.
-        const next = new Map(prev);
-        next.set(cardType.card_type_id, cardType);
-        return next;
-      });
-    };
-    const handleCardTypePatched = (cardType: CardType) => {
-      setCardTypeById((prev) => replaceIfChanged(prev, cardType.card_type_id, cardType));
-    };
-    const handleCardTypeRemoved = (cardType: CardType) => {
-      setCardTypeById((prev) => {
-        if (!prev.has(cardType.card_type_id)) return prev;
-        const next = new Map(prev);
-        next.delete(cardType.card_type_id);
-        return next;
-      });
-    };
-
-    cardTypesService.on('created', handleCardTypeCreated);
-    cardTypesService.on('patched', handleCardTypePatched);
-    cardTypesService.on('updated', handleCardTypePatched);
-    cardTypesService.on('removed', handleCardTypeRemoved);
+    cardTypesService.on('created', realtime.cardTypeCreated);
+    cardTypesService.on('patched', realtime.cardTypePatched);
+    cardTypesService.on('updated', realtime.cardTypePatched);
+    cardTypesService.on('removed', realtime.cardTypeRemoved);
 
     // Subscribe to artifact events
     const artifactsService = client.service('artifacts');
-    const handleArtifactCreated = (artifact: Artifact) => {
-      bumpRevision('artifacts');
-      setArtifactById((prev) => {
-        if (prev.has(artifact.artifact_id)) return prev;
-        const next = new Map(prev);
-        next.set(artifact.artifact_id, artifact);
-        return next;
-      });
-    };
-    const handleArtifactPatched = (artifact: Artifact) => {
-      bumpRevision('artifacts');
-      setArtifactById((prev) => replaceIfChanged(prev, artifact.artifact_id, artifact));
-      // Notify ArtifactNode components that payload may have changed. The
-      // consumer (apps/agor-ui/src/components/SessionCanvas/canvas/ArtifactNode.tsx)
-      // already filters by `contentHash !== lastHashRef.current`, so an
-      // idempotent dispatch is a cheap no-op there — no need to mirror the
-      // shallow-equal bailout from a state-updater side effect (which would
-      // not be pure under StrictMode anyway).
-      window.dispatchEvent(
-        new CustomEvent('agor:artifact-patched', {
-          detail: { artifactId: artifact.artifact_id, contentHash: artifact.content_hash },
-        })
-      );
-    };
-    const handleArtifactRemoved = (artifact: Artifact) => {
-      bumpRevision('artifacts');
-      setArtifactById((prev) => {
-        if (!prev.has(artifact.artifact_id)) return prev;
-        const next = new Map(prev);
-        next.delete(artifact.artifact_id);
-        return next;
-      });
-    };
-
-    artifactsService.on('created', handleArtifactCreated);
-    artifactsService.on('patched', handleArtifactPatched);
-    artifactsService.on('updated', handleArtifactPatched);
-    artifactsService.on('removed', handleArtifactRemoved);
+    artifactsService.on('created', realtime.artifactCreated);
+    artifactsService.on('patched', realtime.artifactPatched);
+    artifactsService.on('updated', realtime.artifactPatched);
+    artifactsService.on('removed', realtime.artifactRemoved);
 
     // Agent-driven runtime queries: daemon emits when an MCP tool wants to
     // introspect the iframe DOM. ArtifactNode components listen for the
@@ -2076,76 +1115,15 @@ export function useAgorData(
 
     // Subscribe to session-MCP server relationship events
     const sessionMcpService = client.service('session-mcp-servers');
-    const handleSessionMcpCreated = (relationship: {
-      session_id: string;
-      mcp_server_id: string;
-    }) => {
-      bumpRevision('sessionMcp');
-      setSessionMcpServerIds((prev) => {
-        const sessionMcpIds = prev.get(relationship.session_id) || [];
-        // Check if relationship already exists (duplicate event)
-        if (sessionMcpIds.includes(relationship.mcp_server_id)) return prev;
-
-        const next = new Map(prev);
-        next.set(relationship.session_id, [...sessionMcpIds, relationship.mcp_server_id]);
-        return next;
-      });
-    };
-    const handleSessionMcpRemoved = (relationship: {
-      session_id: string;
-      mcp_server_id: string;
-    }) => {
-      bumpRevision('sessionMcp');
-      setSessionMcpServerIds((prev) => {
-        const sessionMcpIds = prev.get(relationship.session_id) || [];
-        const filtered = sessionMcpIds.filter((id) => id !== relationship.mcp_server_id);
-
-        // No change if MCP server wasn't in the list
-        if (filtered.length === sessionMcpIds.length) return prev;
-
-        const next = new Map(prev);
-        if (filtered.length > 0) {
-          next.set(relationship.session_id, filtered);
-        } else {
-          // Clean up empty arrays
-          next.delete(relationship.session_id);
-        }
-        return next;
-      });
-    };
-
-    sessionMcpService.on('created', handleSessionMcpCreated);
-    sessionMcpService.on('removed', handleSessionMcpRemoved);
+    sessionMcpService.on('created', realtime.sessionMcpCreated);
+    sessionMcpService.on('removed', realtime.sessionMcpRemoved);
 
     // Subscribe to board comment events
     const commentsService = client.service('board-comments');
-    const handleCommentCreated = (comment: BoardComment) => {
-      bumpRevision('comments');
-      setCommentById((prev) => {
-        if (prev.has(comment.comment_id)) return prev; // Already exists, shouldn't happen
-        const next = new Map(prev);
-        next.set(comment.comment_id, comment);
-        return next;
-      });
-    };
-    const handleCommentPatched = (comment: BoardComment) => {
-      bumpRevision('comments');
-      setCommentById((prev) => replaceIfChanged(prev, comment.comment_id, comment));
-    };
-    const handleCommentRemoved = (comment: BoardComment) => {
-      bumpRevision('comments');
-      setCommentById((prev) => {
-        if (!prev.has(comment.comment_id)) return prev; // Doesn't exist, nothing to remove
-        const next = new Map(prev);
-        next.delete(comment.comment_id);
-        return next;
-      });
-    };
-
-    commentsService.on('created', handleCommentCreated);
-    commentsService.on('patched', handleCommentPatched);
-    commentsService.on('updated', handleCommentPatched);
-    commentsService.on('removed', handleCommentRemoved);
+    commentsService.on('created', realtime.commentCreated);
+    commentsService.on('patched', realtime.commentPatched);
+    commentsService.on('updated', realtime.commentPatched);
+    commentsService.on('removed', realtime.commentRemoved);
 
     // Listen for OAuth completion events to update per-user token state in real-time.
     // Only update the per-user set when oauth_mode is 'per_user' (or unset, which defaults
@@ -2164,7 +1142,7 @@ export function useAgorData(
       bumpRevision('oauth');
       const mode = event.oauth_mode || 'per_user';
       if (mode === 'per_user') {
-        setUserAuthenticatedMcpServerIds((prev) => {
+        agorStore.getState().setMap('userAuthenticatedMcpServerIds', (prev) => {
           if (prev.has(event.mcp_server_id!)) return prev;
           const next = new Set(prev);
           next.add(event.mcp_server_id!);
@@ -2182,7 +1160,9 @@ export function useAgorData(
       try {
         const fresh = (await client.service('mcp-servers').get(event.mcp_server_id)) as MCPServer;
         bumpRevision('mcpServers');
-        setMcpServerById((prev) => replaceIfChanged(prev, fresh.mcp_server_id, fresh));
+        agorStore
+          .getState()
+          .setMap('mcpServerById', (prev) => replaceIfChanged(prev, fresh.mcp_server_id, fresh));
       } catch (err) {
         console.warn('[OAuth] Failed to refetch MCP server after re-auth:', err);
       }
@@ -2197,7 +1177,7 @@ export function useAgorData(
       if (!event.mcp_server_id) return;
       bumpRevision('oauth');
       bumpRevision('mcpServers');
-      setUserAuthenticatedMcpServerIds((prev) => {
+      agorStore.getState().setMap('userAuthenticatedMcpServerIds', (prev) => {
         if (!prev.has(event.mcp_server_id)) return prev;
         const next = new Set(prev);
         next.delete(event.mcp_server_id);
@@ -2210,7 +1190,7 @@ export function useAgorData(
       // `userAuthenticatedMcpServerIds` check — and for tokens with no
       // expiry (e.g. Notion), `isExpired` is always false, so the pill
       // stays purple forever even though the Set was updated above.
-      setMcpServerById((prev) => {
+      agorStore.getState().setMap('mcpServerById', (prev) => {
         const existing = prev.get(event.mcp_server_id);
         if (!existing?.auth?.oauth_access_token) return prev;
         const next = new Map(prev);
@@ -2228,7 +1208,9 @@ export function useAgorData(
       // Still refetch to get the canonical server state from the daemon.
       try {
         const fresh = (await client.service('mcp-servers').get(event.mcp_server_id)) as MCPServer;
-        setMcpServerById((prev) => replaceIfChanged(prev, fresh.mcp_server_id, fresh));
+        agorStore
+          .getState()
+          .setMap('mcpServerById', (prev) => replaceIfChanged(prev, fresh.mcp_server_id, fresh));
       } catch (err) {
         console.warn('[OAuth] Failed to refetch MCP server after disconnect:', err);
       }
@@ -2289,68 +1271,68 @@ export function useAgorData(
       client.io.off('oauth:disconnected', handleOAuthDisconnected);
       client.io.off('connect', refetchSilently);
       window.removeEventListener(TOKENS_REFRESHED_EVENT, handleTokensRefreshed);
-      sessionsService.removeListener('created', handleSessionCreated);
-      sessionsService.removeListener('patched', handleSessionPatched);
-      sessionsService.removeListener('updated', handleSessionPatched);
-      sessionsService.removeListener('removed', handleSessionRemoved);
+      sessionsService.removeListener('created', realtime.sessionCreated);
+      sessionsService.removeListener('patched', realtime.sessionPatched);
+      sessionsService.removeListener('updated', realtime.sessionPatched);
+      sessionsService.removeListener('removed', realtime.sessionRemoved);
 
-      boardsService.removeListener('created', handleBoardCreated);
-      boardsService.removeListener('patched', handleBoardPatched);
-      boardsService.removeListener('updated', handleBoardPatched);
-      boardsService.removeListener('removed', handleBoardRemoved);
+      boardsService.removeListener('created', realtime.boardCreated);
+      boardsService.removeListener('patched', realtime.boardPatched);
+      boardsService.removeListener('updated', realtime.boardPatched);
+      boardsService.removeListener('removed', realtime.boardRemoved);
 
-      boardObjectsService.removeListener('created', handleBoardObjectCreated);
-      boardObjectsService.removeListener('patched', handleBoardObjectPatched);
-      boardObjectsService.removeListener('updated', handleBoardObjectPatched);
-      boardObjectsService.removeListener('removed', handleBoardObjectRemoved);
+      boardObjectsService.removeListener('created', realtime.boardObjectCreated);
+      boardObjectsService.removeListener('patched', realtime.boardObjectPatched);
+      boardObjectsService.removeListener('updated', realtime.boardObjectPatched);
+      boardObjectsService.removeListener('removed', realtime.boardObjectRemoved);
 
-      reposService.removeListener('created', handleRepoCreated);
-      reposService.removeListener('patched', handleRepoPatched);
-      reposService.removeListener('updated', handleRepoPatched);
-      reposService.removeListener('removed', handleRepoRemoved);
+      reposService.removeListener('created', realtime.repoCreated);
+      reposService.removeListener('patched', realtime.repoPatched);
+      reposService.removeListener('updated', realtime.repoPatched);
+      reposService.removeListener('removed', realtime.repoRemoved);
 
-      branchesService.removeListener('created', handleBranchCreated);
-      branchesService.removeListener('patched', handleBranchPatched);
-      branchesService.removeListener('updated', handleBranchPatched);
-      branchesService.removeListener('removed', handleBranchRemoved);
+      branchesService.removeListener('created', realtime.branchCreated);
+      branchesService.removeListener('patched', realtime.branchPatched);
+      branchesService.removeListener('updated', realtime.branchPatched);
+      branchesService.removeListener('removed', realtime.branchRemoved);
 
-      usersService.removeListener('created', handleUserCreated);
-      usersService.removeListener('patched', handleUserPatched);
-      usersService.removeListener('updated', handleUserPatched);
-      usersService.removeListener('removed', handleUserRemoved);
+      usersService.removeListener('created', realtime.userCreated);
+      usersService.removeListener('patched', realtime.userPatched);
+      usersService.removeListener('updated', realtime.userPatched);
+      usersService.removeListener('removed', realtime.userRemoved);
 
-      mcpServersService.removeListener('created', handleMCPServerCreated);
-      mcpServersService.removeListener('patched', handleMCPServerPatched);
-      mcpServersService.removeListener('updated', handleMCPServerPatched);
-      mcpServersService.removeListener('removed', handleMCPServerRemoved);
+      mcpServersService.removeListener('created', realtime.mcpServerCreated);
+      mcpServersService.removeListener('patched', realtime.mcpServerPatched);
+      mcpServersService.removeListener('updated', realtime.mcpServerPatched);
+      mcpServersService.removeListener('removed', realtime.mcpServerRemoved);
 
-      sessionMcpService.removeListener('created', handleSessionMcpCreated);
-      sessionMcpService.removeListener('removed', handleSessionMcpRemoved);
+      sessionMcpService.removeListener('created', realtime.sessionMcpCreated);
+      sessionMcpService.removeListener('removed', realtime.sessionMcpRemoved);
 
-      commentsService.removeListener('created', handleCommentCreated);
-      commentsService.removeListener('patched', handleCommentPatched);
-      commentsService.removeListener('updated', handleCommentPatched);
-      commentsService.removeListener('removed', handleCommentRemoved);
+      commentsService.removeListener('created', realtime.commentCreated);
+      commentsService.removeListener('patched', realtime.commentPatched);
+      commentsService.removeListener('updated', realtime.commentPatched);
+      commentsService.removeListener('removed', realtime.commentRemoved);
 
-      gatewayChannelsService.removeListener('created', handleGatewayChannelCreated);
-      gatewayChannelsService.removeListener('patched', handleGatewayChannelPatched);
-      gatewayChannelsService.removeListener('updated', handleGatewayChannelPatched);
-      gatewayChannelsService.removeListener('removed', handleGatewayChannelRemoved);
+      gatewayChannelsService.removeListener('created', realtime.gatewayChannelCreated);
+      gatewayChannelsService.removeListener('patched', realtime.gatewayChannelPatched);
+      gatewayChannelsService.removeListener('updated', realtime.gatewayChannelPatched);
+      gatewayChannelsService.removeListener('removed', realtime.gatewayChannelRemoved);
 
-      cardsService.removeListener('created', handleCardCreated);
-      cardsService.removeListener('patched', handleCardPatched);
-      cardsService.removeListener('updated', handleCardPatched);
-      cardsService.removeListener('removed', handleCardRemoved);
+      cardsService.removeListener('created', realtime.cardCreated);
+      cardsService.removeListener('patched', realtime.cardPatched);
+      cardsService.removeListener('updated', realtime.cardPatched);
+      cardsService.removeListener('removed', realtime.cardRemoved);
 
-      cardTypesService.removeListener('created', handleCardTypeCreated);
-      cardTypesService.removeListener('patched', handleCardTypePatched);
-      cardTypesService.removeListener('updated', handleCardTypePatched);
-      cardTypesService.removeListener('removed', handleCardTypeRemoved);
+      cardTypesService.removeListener('created', realtime.cardTypeCreated);
+      cardTypesService.removeListener('patched', realtime.cardTypePatched);
+      cardTypesService.removeListener('updated', realtime.cardTypePatched);
+      cardTypesService.removeListener('removed', realtime.cardTypeRemoved);
 
-      artifactsService.removeListener('created', handleArtifactCreated);
-      artifactsService.removeListener('patched', handleArtifactPatched);
-      artifactsService.removeListener('updated', handleArtifactPatched);
-      artifactsService.removeListener('removed', handleArtifactRemoved);
+      artifactsService.removeListener('created', realtime.artifactCreated);
+      artifactsService.removeListener('patched', realtime.artifactPatched);
+      artifactsService.removeListener('updated', realtime.artifactPatched);
+      artifactsService.removeListener('removed', realtime.artifactRemoved);
       artifactsService.removeListener('agor-query', handleAgorQuery);
     };
   }, [client, enabled, fetchData, hasInitiallyFetched]);
@@ -2360,21 +1342,23 @@ export function useAgorData(
   const initialLoadItems = useMemo<InitialLoadItem[]>(
     () =>
       INITIAL_LOAD_ITEMS.map(({ key, label }) => {
-        const count = itemCounts[key];
+        const count = storeState.itemCounts[key];
         return { key, label, done: count !== undefined, count: count ?? 0 };
       }),
-    [itemCounts]
+    [storeState.itemCounts]
   );
 
-  const initialLoadComplete = INITIAL_LOAD_ITEMS.every(({ key }) => itemCounts[key] !== undefined);
+  const initialLoadComplete = INITIAL_LOAD_ITEMS.every(
+    ({ key }) => storeState.itemCounts[key] !== undefined
+  );
 
   return {
-    ...maps,
+    ...pickMaps(storeState),
     initialLoadItems,
     initialLoadComplete,
-    loadingStage,
-    loading,
-    error,
+    loadingStage: storeState.loadingStage,
+    loading: storeState.loading,
+    error: storeState.error,
     refetch: fetchData,
   };
 }
